fix: adhere to shellcheck and update yadm

+++ scripts/add_project.sh
@@ -10,9 +10,9 @@ get_submodule_name() {
         exit 1
     fi
     # Read repository name from parent directories remote url
-    SUBMODULE_NAME="$(basename -s ".git" $PARENT_REPOSITORY_REMOTE)-dotfiles"
+    SUBMODULE_NAME="$(basename -s ".git" "$PARENT_REPOSITORY_REMOTE")-dotfiles"
     
-    read -p $'\nPlease provide the name of the project\n(default will append "-dotfiles" to the parent repositories name )\n' -e -i "$SUBMODULE_NAME" SUBMODULE_NAME
+    read -er -p $'\nPlease provide the name of the project\n(default will append "-dotfiles" to the parent repositories name )\n' -i "$SUBMODULE_NAME" SUBMODULE_NAME
 }
 
 make_new_folder() {
@@ -31,32 +31,32 @@ make_new_folder() {
 initialize_submodule_repo() {
     SUBMODULE_DIR="$PWD/.secrets"
     # Prompt for git remote with default parsed from parent repository
-    read -p $'\nProvide a git remote to store this new module in:\n' -e -i "$(dirname $PARENT_REPOSITORY_REMOTE)/$SUBMODULE_NAME.git" REMOTE_URL
+    read -er -p $'\nProvide a git remote to store this new module in:\n' -i "$(dirname "$PARENT_REPOSITORY_REMOTE")/$SUBMODULE_NAME.git" REMOTE_URL
     # Prompt for git default branch with default parsed from git config
-    read -p $'\nProvide a default branch name\n' -e -i "$(git config --get init.defaultBranch || echo "master")" BRANCH
-    if [ $BRANCH == "" ]; then
+    read -er -p $'\nProvide a default branch name\n' -i "$(git config --get init.defaultBranch || echo "master")" BRANCH
+    if [ "$BRANCH" == "" ]; then
         echo "Invalid Branch Name"
         exit 1
     fi
 
     # TODO: add in step to go create the repo in github/gitlab
     # TODO: Make the repository created here private by default, Look at this gist for inspiration for github https://gist.github.com/alexpchin/dc91e723d4db5018fef8
-    URL_ENCODED_REPO_NAME=$(jq -rn --arg x $SUBMODULE_NAME '$x|@uri')
-    read -p "Please follow this link to create a !!!PRIVATE!!! repository with the correct name of $SUBMODULE_NAME"$'\n'"https://github.com/new?name=$URL_ENCODED_REPO_NAME"$'\nPress enter when ready...'
+    URL_ENCODED_REPO_NAME=$(jq -rn --arg x "$SUBMODULE_NAME" '$x|@uri')
+    read -r -p "Please follow this link to create a !!!PRIVATE!!! repository with the correct name of $SUBMODULE_NAME"$'\n'"https://github.com/new?name=$URL_ENCODED_REPO_NAME"$'\nPress enter when ready...'
 
-    echo "# $SUBMODULE_NAME" >> $SUBMODULE_NAME/README.md
-    git -C $SUBMODULE_DIR init
-    git -C $SUBMODULE_DIR add README.md
-    git -C $SUBMODULE_DIR commit -m "feat: initial commit"
-    git -C $SUBMODULE_DIR branch -M $BRANCH
-    git -C $SUBMODULE_DIR remote add origin $REMOTE_URL
-    git -C $SUBMODULE_DIR push -u origin $BRANCH
+    echo "# $SUBMODULE_NAME" >> "$SUBMODULE_NAME/README.md"
+    git -C "$SUBMODULE_DIR" init
+    git -C "$SUBMODULE_DIR" add README.md
+    git -C "$SUBMODULE_DIR" commit -m "feat: initial commit"
+    git -C "$SUBMODULE_DIR" branch -M "$BRANCH"
+    git -C "$SUBMODULE_DIR" remote add origin "$REMOTE_URL"
+    git -C "$SUBMODULE_DIR" push -u origin "$BRANCH"
     # Remove newly created repository to make room for submodule
-    rm -rf "$PWD/$SUBMODULE_NAME"
-    git submodule add $REMOTE_URL
+    rm -rf "$PWD/${SUBMODULE_NAME:?}"
+    git submodule add "$REMOTE_URL"
 }
 
-source $PWD/scripts/install_package.sh jq
+source "./install_package.sh" jq
 get_submodule_name
 make_new_folder
 initialize_submodule_repo
